refactor(frontend): extract git pane and improve sidebar layout
- Extract GitPane component into a separate file for better organization
- Add scrollbar styling to improve sidebar appearance
- Simplify LeftSidebar component by removing Git-related logic
- Implement cleaner git status and commit interface
